Add shebangs
Executable scripts need a shebang to find an interpreter to run the
script with. These scripts do not have a shebang and would need to be
used as 'bash xtensa-trace.sh' (if executing with any other shell).
However, most of the docs and generally people would execute it as
'./xtensa-trace.sh' which needs shebangs.

Signed-off-by: Shreeya Patel <shreeya.patel23498@gmail.com>